refactor: #16 게시글 미디어 서비스 제거 및 로직 조정
